# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TogoMarket API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings/{listingId}/reviews:
    get:
      operationId: getListingReviews
      tags: [reviews]
      summary: Get reviews for a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews for the listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsList"
    post:
      operationId: createListingReview
      tags: [reviews]
      summary: Create a review on a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created (editToken returned once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCreateResult"
        "404":
          description: Listing not found
  /reviews/update:
    post:
      operationId: updateReview
      tags: [reviews]
      summary: Update own review (author only, via editToken)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdateInput"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "403":
          description: Not the author
        "404":
          description: Review not found
  /reviews/delete:
    post:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (author via editToken, or admin via password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDeleteInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Not allowed
        "404":
          description: Review not found
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    post:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
        "403":
          description: Forbidden
  /admin/sub-admin-verify:
    post:
      operationId: verifySubAdmin
      tags: [admin]
      summary: Verify sub-admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubAdminVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubAdminVerifyResponse"
...
  /admin/services:
    post:
      operationId: adminCreateService
      tags: [admin, services]
      summary: Create a new service offer (admin/sub-admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
  /admin/services/all:
    post:
      operationId: adminGetAllServices
      tags: [admin, services]
      summary: Get all services including expired (admin/sub-admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPasswordInput"
      responses:
        "200":
          description: All services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
        "403":
          description: Forbidden
  /admin/services/delete:
    post:
      operationId: adminDeleteService
      tags: [admin, services]
      summary: Delete a service (admin/sub-admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDeleteInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Review:
      type: object
      required: [id, listingId, buyerName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [buyerName, buyerPhone, rating]
      properties:
        buyerName:
          type: string
          minLength: 2
        buyerPhone:
          type: string
          minLength: 8
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    ReviewCreateResult:
      type: object
      required: [review, editToken]
      properties:
        review:
          $ref: "#/components/schemas/Review"
        editToken:
          type: string
    ReviewUpdateInput:
      type: object
      required: [id, editToken, rating]
      properties:
        id:
          type: integer
        editToken:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    ReviewDeleteInput:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        editToken:
          type: string
        password:
          type: string
    ReviewsList:
      type: object
      required: [items, avgRating, count]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        avgRating:
          type: ["number", "null"]
        count:
          type: integer
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    AdminDeleteInput:
      type: object
      required: [id, password]
      properties:
        id:
          type: integer
        password:
          type: string
...
    AdminPasswordInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    PlatformSettings:
      type: object
      required: [commissionRate, whatsappCommission, whatsappOrders, whatsappAds, whatsappServices, otpProvider, whatsappValidation]
      properties:
        commissionRate:
          type: integer
        whatsappCommission:
          type: string
        whatsappOrders:
          type: string
        whatsappAds:
          type: string
        whatsappServices:
          type: string
        otpProvider:
          type: string
          enum: [WHATSAPP, TECHSOFT, MANUAL]
        whatsappValidation:
          type: string
    UpdateSettingsInput:
      type: object
      required: [password, commissionRate, whatsappCommission, whatsappOrders]
      properties:
        password:
          type: string
        commissionRate:
          type: integer
        whatsappCommission:
          type: string
        whatsappOrders:
          type: string
        subAdminPassword:
          type: string
        whatsappAds:
          type: string
        whatsappServices:
          type: string
        otpProvider:
          type: string
          enum: [WHATSAPP, TECHSOFT, MANUAL]
        whatsappValidation:
          type: string
    Service:
      type: object
      required: [id, type, title, description, contact, quartier, ville, createdAt, expiresAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        contact:
          type: string
        quartier:
          type: string
        ville:
          type: string
        image:
          type: string
          nullable: true
        videoPath:
          type: ["string", "null"]
        createdAt:
          type: string
        expiresAt:
          type: string
        isPublished:
          type: ["boolean", "null"]
    AdminCreateServiceInput:
      type: object
      required: [password, type, title, description, contact, quartier, ville]
      properties:
        password:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        contact:
          type: string
        quartier:
          type: string
        ville:
          type: string
        image:
          type: string
        videoPath:
          type: string
    SubAdminVerifyInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    SubAdminVerifyResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
...