# lib/api-zod/src/generated/types/platformSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TogoMarket API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformSettingsOtpProvider } from './platformSettingsOtpProvider';

export interface PlatformSettings {
  commissionRate: number;
  whatsappCommission: string;
  whatsappOrders: string;
  whatsappAds: string;
  whatsappServices: string;
  otpProvider: PlatformSettingsOtpProvider;
  whatsappValidation: string;
}