# lib/api-zod/src/generated/types/updateSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TogoMarket API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSettingsInputOtpProvider } from './updateSettingsInputOtpProvider';

export interface UpdateSettingsInput {
  password: string;
  commissionRate: number;
  whatsappCommission: string;
  whatsappOrders: string;
  subAdminPassword?: string;
  whatsappAds?: string;
  whatsappServices?: string;
  otpProvider?: UpdateSettingsInputOtpProvider;
  whatsappValidation?: string;
}